basic user authentication and author status

+++ posts/views.py
@@ -1,7 +1,7 @@
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
-from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.shortcuts import render, get_object_or_404, redirect
 
 from urllib.parse import quote_plus
 
@@ -12,10 +12,15 @@ from .forms import PostForm
 # Function based views vs class based views
 
 def post_create(request):
+    # if not request.user.is_staff or not request.user.is_superuser:
+    #     raise Http404
+    if not request.user.is_authenticated:
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instance = form.save(commit=False)
-        print(form.cleaned_data.get("title"))
+        instance.user = request.user
+        # print(form.cleaned_data.get("title"))
         instance.save()
         messages.success(request, "Successfully Created")
         return HttpResponseRedirect(instance.get_absolute_url())
@@ -51,6 +56,8 @@ def post_list(request):
 
 
 def post_update(request, slug=None):
+    if not request.user.is_authenticated:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -67,6 +74,8 @@ def post_update(request, slug=None):
 
 
 def post_delete(request, slug=None):
+    if not request.user.is_authenticated:
+        raise Http404
     instance = get_object_or_404(Post, slug=slug)
     instance.delete()
     messages.success(request, "Successfully Deleted Blog Entry")
